+++ tipg/collections.py
@@ -412,7 +412,6 @@ class PgCollection(Collection):
         tms: TileMatrixSet,
         tile: Tile,
         simplify: Optional[float],
-        preserve_topology: Optional[bool],
     ):
         """Create MVT from intersecting geometries."""
         geom = pg_funcs.cast(logic.V(geometry_column.name), "geometry")
@@ -420,7 +419,7 @@ class PgCollection(Collection):
         if simplify:
             geom = logic.Func(
                 "ST_SnapToGrid",
-                logic.Func("ST_SimplifyPreserveTopology" if preserve_topology else "ST_Simplify", geom, simplify),
+                logic.Func("ST_SimplifyPreserveTopology", geom, simplify),
                 simplify,
             )
 
@@ -876,7 +875,6 @@ class PgCollection(Collection):
         dt: Optional[str] = None,
         limit: Optional[int] = None,
         simplify: Optional[float] = None,
-        preserve_topology: Optional[bool] = None,
     ):
         """Build query to get Vector Tile."""
         limit = limit or mvt_settings.max_features_per_tile
@@ -897,7 +895,6 @@ class PgCollection(Collection):
                 tms=tms,
                 tile=tile,
                 simplify=simplify,
-                preserve_topology=preserve_topology,
             ),
             self._from(function_parameters),
             self._where(