+++ tipg/collections.py
@@ -411,9 +411,17 @@ class PgCollection(Collection):
         geometry_column: Column,
         tms: TileMatrixSet,
         tile: Tile,
+        simplify: Optional[float],
     ):
         """Create MVT from intersecting geometries."""
         geom = pg_funcs.cast(logic.V(geometry_column.name), "geometry")
+
+        if simplify:
+            geom = logic.Func(
+                "ST_SnapToGrid",
+                logic.Func("ST_SimplifyPreserveTopology", geom, simplify),
+                simplify,
+            )
 
         # make sure the geometries do not overflow the TMS bbox
         if not tms.is_valid(tile):
@@ -866,6 +874,7 @@ class PgCollection(Collection):
         geom: Optional[str] = None,
         dt: Optional[str] = None,
         limit: Optional[int] = None,
+        simplify: Optional[float] = None,
     ):
         """Build query to get Vector Tile."""
         limit = limit or mvt_settings.max_features_per_tile
@@ -885,6 +894,7 @@ class PgCollection(Collection):
                 geometry_column=geometry_column,
                 tms=tms,
                 tile=tile,
+                simplify=simplify,
             ),
             self._from(function_parameters),
             self._where(